+++ src/logica/Calculadora.java
@@ -68,12 +68,12 @@ public class Calculadora
 		}
 	}
 	
-	public int realizarOperacion(String nombre, int param1, int param2) throws PluginException, OperacionException
+	public float realizarOperacion(String nombre, int param1, int param2) throws PluginException, OperacionException
 	{
 		if(plugins.size()==0)
 			throw new PluginException("No hay plugins cargados.");
 		
-		int toReturn;
+		float toReturn;
 		PluginInterface plugin = getOperacion(nombre);
 		plugin.setParametros(param1, param2);
